backend/requirements.txt
# Django and REST Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Message Queue
pika==1.3.2

# API and HTTP
requests==2.31.0
httpx==0.26.0

# Utilities
python-decouple==3.8
celery==5.3.4
redis==5.0.1

# Development and debugging
django-extensions==3.2.3
ipython==8.18.1

# Validation and serialization
marshmallow==3.20.2

# Pagination
django-rest-framework-pagination==0.1.1

# Documentation (optional)
drf-spectacular==0.27.0